requirements.txt
annotated-types==0.7.0
anyio==4.6.0
beautifulsoup4==4.12.3
certifi==2024.7.4
charset-normalizer==3.3.2
distro==1.9.0
exceptiongroup==1.2.2
feedparser==6.0.11
frozendict==2.4.4
groq==0.11.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httpx==0.27.2
idna==3.7
inflection==0.5.1
lxml==5.2.2
more-itertools==10.3.0
multitasking==0.0.11
Nasdaq-Data-Link==1.0.4
numpy==2.0.1
pandas==2.2.2
peewee==3.17.6
platformdirs==4.2.2
pydantic==2.9.2
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
schedule==1.2.2
sgmllib3k==1.0.0
six==1.16.0
slack_sdk==3.31.0
sniffio==1.3.1
soupsieve==2.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
webencodings==0.5.1
yfinance==0.2.41
